third_party/WebKit/LayoutTests/fast/spatial-navigation/snav-symmetrically-positioned-expected.txt

This test ensures that Spatial Navigation works with symmetrically positoned rects

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


E1E2E3E4E5
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e1"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e3"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "e5"
PASS successfullyParsed is true

TEST COMPLETE
